More reformatting and automated code-review changes.
