+++ samples/Jcd.Utilities.Samples.Console/Generators/BigIntegerSequenceState.cs
@@ -1,0 +1,6 @@
+﻿using System.Numerics;
+
+namespace Jcd.Utilities.Samples.ConsoleApp.Generators
+{
+    public class BigIntegerSequenceState : EquatableSequenceState<BigInteger> { }
+}
