+++ samples/Jcd.Utilities.Samples.Console/Generators/ByteSequenceState.cs
@@ -1,0 +1,4 @@
+﻿namespace Jcd.Utilities.Samples.ConsoleApp.Generators
+{
+    public class ByteSequenceState : EquatableSequenceState<byte> { }
+}
