+++ samples/Jcd.Utilities.Samples.Console/Generators/Int16SequenceState.cs
@@ -1,0 +1,4 @@
+﻿namespace Jcd.Utilities.Samples.ConsoleApp.Generators
+{
+    public class Int16SequenceState : EquatableSequenceState<short> { }
+}
