+++ samples/Jcd.Utilities.Samples.Console/Generators/Int32SequenceState.cs
@@ -1,0 +1,4 @@
+﻿namespace Jcd.Utilities.Samples.ConsoleApp.Generators
+{
+    public class Int32SequenceState : EquatableSequenceState<int> { }
+}
