+++ samples/Jcd.Utilities.Samples.Console/Generators/Int64SequenceState.cs
@@ -1,0 +1,4 @@
+﻿namespace Jcd.Utilities.Samples.ConsoleApp.Generators
+{
+    public class Int64SequenceState : EquatableSequenceState<long> { }
+}
