--- samples/Jcd.Utilities.Samples.Console/Generators/IntSequenceState.cs
@@ -1,22 +1,0 @@
-﻿using System.Numerics;
-
-namespace Jcd.Utilities.Samples.ConsoleApp.Generators
-{
-    public class BigIntegerSequenceState : EquatableSequenceState<BigInteger> { }
-
-    public class Int64SequenceState : EquatableSequenceState<long> { }
-
-    public class Int32SequenceState : EquatableSequenceState<int> { }
-
-    public class Int16SequenceState : EquatableSequenceState<short> { }
-
-    public class SByteSequenceState : EquatableSequenceState<sbyte> { }
-
-    public class UInt64SequenceState : EquatableSequenceState<ulong> { }
-
-    public class UInt32SequenceState : EquatableSequenceState<uint> { }
-
-    public class UInt16SequenceState : EquatableSequenceState<ushort> { }
-
-    public class ByteSequenceState : EquatableSequenceState<byte> { }
-}
