+++ samples/Jcd.Utilities.Samples.Console/Generators/SByteSequenceState.cs
@@ -1,0 +1,4 @@
+﻿namespace Jcd.Utilities.Samples.ConsoleApp.Generators
+{
+    public class SByteSequenceState : EquatableSequenceState<sbyte> { }
+}
