+++ samples/Jcd.Utilities.Samples.Console/Generators/UInt16SequenceState.cs
@@ -1,0 +1,4 @@
+﻿namespace Jcd.Utilities.Samples.ConsoleApp.Generators
+{
+    public class UInt16SequenceState : EquatableSequenceState<ushort> { }
+}
