+++ samples/Jcd.Utilities.Samples.Console/Generators/UInt32SequenceState.cs
@@ -1,0 +1,4 @@
+﻿namespace Jcd.Utilities.Samples.ConsoleApp.Generators
+{
+    public class UInt32SequenceState : EquatableSequenceState<uint> { }
+}
