+++ samples/Jcd.Utilities.Samples.Console/Generators/UInt64SequenceState.cs
@@ -1,0 +1,4 @@
+﻿namespace Jcd.Utilities.Samples.ConsoleApp.Generators
+{
+    public class UInt64SequenceState : EquatableSequenceState<ulong> { }
+}
