+++ src/Jcd.Utilities/Formatting/IIntegerParser.cs
@@ -7,7 +7,11 @@ namespace Jcd.Utilities
     /// </summary>
     public interface IIntegerParser
     {
-        sbyte ParseSByte(string value);
+        #region Public Methods
+
+        BigInteger ParseBigInteger(string value);
+
+        byte ParseByte(string value);
 
         short ParseInt16(string value);
 
@@ -15,9 +19,7 @@ namespace Jcd.Utilities
 
         long ParseInt64(string value);
 
-        BigInteger ParseBigInteger(string value);
-
-        byte ParseByte(string value);
+        sbyte ParseSByte(string value);
 
         ushort ParseUInt16(string value);
 
@@ -25,7 +27,9 @@ namespace Jcd.Utilities
 
         ulong ParseUInt64(string value);
 
-        bool TryParseSByte(string value, ref sbyte result);
+        bool TryParseBigInteger(string value, ref BigInteger result);
+
+        bool TryParseByte(string value, ref byte result);
 
         bool TryParseInt16(string value, ref short result);
 
@@ -33,14 +37,14 @@ namespace Jcd.Utilities
 
         bool TryParseInt64(string value, ref long result);
 
-        bool TryParseBigInteger(string value, ref BigInteger result);
-
-        bool TryParseByte(string value, ref byte result);
+        bool TryParseSByte(string value, ref sbyte result);
 
         bool TryParseUInt16(string value, ref ushort result);
 
         bool TryParseUInt32(string value, ref uint result);
 
         bool TryParseUInt64(string value, ref ulong result);
+
+        #endregion Public Methods
     }
 }
