+++ src/Jcd.Utilities/Validations/Check.cs
@@ -17,6 +17,8 @@ namespace Jcd.Utilities.Validations
     /// </remarks>
     public static class Check
     {
+        #region Public Delegates
+
         /// <summary>
         /// The signature for delegates used in Passes&lt; <typeparamref name="T"/>&gt;, Fails&lt;
         /// <typeparamref name="T"/>&gt;, and aggregates of these methods (e.g. PassesAll). Most
@@ -33,23 +35,9 @@ namespace Jcd.Utilities.Validations
         /// <returns>true if successful, false otherwise.</returns>
         public delegate bool Signature<T>(T value, Action onSuccess = null, Action onFailure = null);
 
-        #region Boolean and Null checks
+        #endregion Public Delegates
 
-        /// <summary>
-        /// Returns the value of value, and executes any success or failure conditions.
-        /// </summary>
-        /// <remarks>
-        /// This is a rudimentary helper method. While public it has little value outside of this or
-        /// a similar framework.
-        /// </remarks>
-        /// <param name="value">The value to evaluate and return</param>
-        /// <param name="onSuccess">The action to take, if any, when value == true.</param>
-        /// <param name="onFailure">The action to take, if any, when value == true.</param>
-        /// <returns>the value of value</returns>
-        public static bool IsTrue(bool value, Action onSuccess = null, Action onFailure = null)
-        {
-            return Passes(() => value, onSuccess, onFailure);
-        }
+        #region Boolean and Null checks
 
         /// <summary>
         /// Returns the logical complement of value, and executes any success or failure conditions.
@@ -68,6 +56,20 @@ namespace Jcd.Utilities.Validations
         }
 
         /// <summary>
+        /// Check if the provided value is not null and, if applicable, take an indicated action.
+        /// </summary>
+        /// <typeparam name="T">The type of the data being evaluated</typeparam>
+        /// <param name="value">The value being evaluated</param>
+        /// <param name="onSuccess">The action to take, if any, when value != null.</param>
+        /// <param name="onFailure">The action to take, if any, when value != null.</param>
+        /// <returns>true if value is null, false otherwise</returns>
+        public static bool IsNotNull<T>(T value, Action onSuccess = null, Action onFailure = null)
+        where T : class
+        {
+            return Passes(() => value != null, onSuccess, onFailure);
+        }
+
+        /// <summary>
         /// Check if the provided value is null and, if applicable, take an indicated action.
         /// </summary>
         /// <typeparam name="T">The type of the data being evaluated</typeparam>
@@ -82,50 +84,24 @@ namespace Jcd.Utilities.Validations
         }
 
         /// <summary>
-        /// Check if the provided value is not null and, if applicable, take an indicated action.
+        /// Returns the value of value, and executes any success or failure conditions.
         /// </summary>
-        /// <typeparam name="T">The type of the data being evaluated</typeparam>
-        /// <param name="value">The value being evaluated</param>
-        /// <param name="onSuccess">The action to take, if any, when value != null.</param>
-        /// <param name="onFailure">The action to take, if any, when value != null.</param>
-        /// <returns>true if value is null, false otherwise</returns>
-        public static bool IsNotNull<T>(T value, Action onSuccess = null, Action onFailure = null)
-        where T : class
+        /// <remarks>
+        /// This is a rudimentary helper method. While public it has little value outside of this or
+        /// a similar framework.
+        /// </remarks>
+        /// <param name="value">The value to evaluate and return</param>
+        /// <param name="onSuccess">The action to take, if any, when value == true.</param>
+        /// <param name="onFailure">The action to take, if any, when value == true.</param>
+        /// <returns>the value of value</returns>
+        public static bool IsTrue(bool value, Action onSuccess = null, Action onFailure = null)
         {
-            return Passes(() => value != null, onSuccess, onFailure);
+            return Passes(() => value, onSuccess, onFailure);
         }
 
         #endregion Boolean and Null checks
 
         #region collection operations
-
-        /// <summary>
-        /// Checks if a collection of type <typeparamref name="T"/> lacks entries.
-        /// </summary>
-        /// <typeparam name="T">The type of the data stored in T</typeparam>
-        /// <param name="list">The collection to check</param>
-        /// <param name="onSuccess">The action to take, if any, when the collection is empty.</param>
-        /// <param name="onFailure">The action to take, if any, when the collection is not empty.</param>
-        /// <returns>True if list is empty, false otherwise</returns>
-        public static bool IsEmpty<T>(IEnumerable<T> list, Action onSuccess = null, Action onFailure = null)
-        {
-            EnforceNonNull(list);
-            return Passes(() => !HasItems(list), onSuccess, onFailure);
-        }
-
-        /// <summary>
-        /// Checks if a collection of type <typeparamref name="T"/> has entries.
-        /// </summary>
-        /// <typeparam name="T">The type of the data stored in T</typeparam>
-        /// <param name="list">The collection to check</param>
-        /// <param name="onSuccess">The action to take, if any, when the collection has entries.</param>
-        /// <param name="onFailure">The action to take, if any, when the collection is empty.</param>
-        /// <returns>False if list is empty, true otherwise</returns>
-        public static bool HasItems<T>(IEnumerable<T> list, Action onSuccess = null, Action onFailure = null)
-        {
-            EnforceNonNull(list);
-            return Passes(() => list.Any(), onSuccess, onFailure);
-        }
 
         /// <summary>
         /// Checks if a collection of type <typeparamref name="T"/> contains a specific item.
@@ -157,21 +133,37 @@ namespace Jcd.Utilities.Validations
             return Passes(() => !Contains(list, target), onSuccess, onFailure);
         }
 
+        /// <summary>
+        /// Checks if a collection of type <typeparamref name="T"/> has entries.
+        /// </summary>
+        /// <typeparam name="T">The type of the data stored in T</typeparam>
+        /// <param name="list">The collection to check</param>
+        /// <param name="onSuccess">The action to take, if any, when the collection has entries.</param>
+        /// <param name="onFailure">The action to take, if any, when the collection is empty.</param>
+        /// <returns>False if list is empty, true otherwise</returns>
+        public static bool HasItems<T>(IEnumerable<T> list, Action onSuccess = null, Action onFailure = null)
+        {
+            EnforceNonNull(list);
+            return Passes(() => list.Any(), onSuccess, onFailure);
+        }
+
+        /// <summary>
+        /// Checks if a collection of type <typeparamref name="T"/> lacks entries.
+        /// </summary>
+        /// <typeparam name="T">The type of the data stored in T</typeparam>
+        /// <param name="list">The collection to check</param>
+        /// <param name="onSuccess">The action to take, if any, when the collection is empty.</param>
+        /// <param name="onFailure">The action to take, if any, when the collection is not empty.</param>
+        /// <returns>True if list is empty, false otherwise</returns>
+        public static bool IsEmpty<T>(IEnumerable<T> list, Action onSuccess = null, Action onFailure = null)
+        {
+            EnforceNonNull(list);
+            return Passes(() => !HasItems(list), onSuccess, onFailure);
+        }
+
         #endregion collection operations
 
         #region string operations
-
-        /// <summary>
-        /// Checks if a string is empty and not-null
-        /// </summary>
-        /// <param name="value">the value to check.</param>
-        /// <param name="onSuccess">The action to take if the string is empty</param>
-        /// <param name="onFailure">The action to take if the string is not empty</param>
-        /// <returns>True if the string is empty, false otherwise.</returns>
-        public static bool IsEmpty(string value, Action onSuccess = null, Action onFailure = null)
-        {
-            return Passes(() => value != null && value.Length == 0, onSuccess, onFailure);
-        }
 
         /// <summary>
         /// Checks if a string has 1 or more characters in it.
@@ -186,17 +178,15 @@ namespace Jcd.Utilities.Validations
         }
 
         /// <summary>
-        /// Checks if the string has only whitespace and is not empty nor null.
+        /// Checks if a string is empty and not-null
         /// </summary>
-        /// <param name="value">The string to check.</param>
-        /// <param name="onSuccess">The action to take if the string is not empty.</param>
-        /// <param name="onFailure">
-        /// The action to take if the string is null, empty, or has nonwhitespace characters.
-        /// </param>
-        /// <returns>True if the string is non-zero length and only contains whitespace.</returns>
-        public static bool IsWhitespace(string value, Action onSuccess = null, Action onFailure = null)
+        /// <param name="value">the value to check.</param>
+        /// <param name="onSuccess">The action to take if the string is empty</param>
+        /// <param name="onFailure">The action to take if the string is not empty</param>
+        /// <returns>True if the string is empty, false otherwise.</returns>
+        public static bool IsEmpty(string value, Action onSuccess = null, Action onFailure = null)
         {
-            return Passes(() => HasData(value) && value.TrimStart() == "", onSuccess, onFailure);
+            return Passes(() => value != null && value.Length == 0, onSuccess, onFailure);
         }
 
         /// <summary>
@@ -213,72 +203,23 @@ namespace Jcd.Utilities.Validations
             return Passes(() => !IsNull(value) && !IsWhitespace(value), onSuccess, onFailure);
         }
 
+        /// <summary>
+        /// Checks if the string has only whitespace and is not empty nor null.
+        /// </summary>
+        /// <param name="value">The string to check.</param>
+        /// <param name="onSuccess">The action to take if the string is not empty.</param>
+        /// <param name="onFailure">
+        /// The action to take if the string is null, empty, or has nonwhitespace characters.
+        /// </param>
+        /// <returns>True if the string is non-zero length and only contains whitespace.</returns>
+        public static bool IsWhitespace(string value, Action onSuccess = null, Action onFailure = null)
+        {
+            return Passes(() => HasData(value) && value.TrimStart() == "", onSuccess, onFailure);
+        }
+
         #endregion string operations
 
         #region range and relational operations
-
-        /// <summary>
-        /// Checks if <paramref name="left"/> and <paramref name="right"/> are the same instance of
-        /// an object.
-        /// </summary>
-        /// <param name="left">The lefthand side of the comparison</param>
-        /// <param name="right">The righthand side of the comparison</param>
-        /// <param name="onSuccess">The action to take if they're the same instance.</param>
-        /// <param name="onFailure">The action to take if they're not the same instance.</param>
-        /// <returns>true if left and right are the same instance.</returns>
-        /// <exception cref="ArgumentNullException">
-        /// If <paramref name="left"/> or <paramref name="right"/> are null.
-        /// </exception>
-        public static bool AreSameObject(object left, object right, Action onSuccess = null, Action onFailure = null)
-        {
-            return Passes(() => ReferenceEquals(left, right), onSuccess, onFailure);
-        }
-
-        /// <summary>
-        /// Checks if <paramref name="left"/> is greater than <paramref name="right"/>.
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="left">The lefthand side of the comparison</param>
-        /// <param name="right">The righthand side of the comparison</param>
-        /// <param name="onSuccess">
-        /// The action to take if <paramref name="left"/> is greater than <paramref name="right"/>
-        /// </param>
-        /// <param name="onFailure">
-        /// The action to take if <paramref name="left"/> is not greater than <paramref name="right"/>
-        /// </param>
-        /// <returns>True if left is greater than right.</returns>
-        /// <exception cref="ArgumentNullException">
-        /// If <paramref name="left"/> or <paramref name="right"/> are null.
-        /// </exception>
-        public static bool IsGreaterThan<T>(T left, T right, Action onSuccess = null, Action onFailure = null)
-        where T : IComparable<T>
-        {
-            EnforceNonNull(left, right);
-            return Passes(() => left.CompareTo(right) > 0, onSuccess, onFailure);
-        }
-
-        /// <summary>
-        /// Checks if <paramref name="left"/> is less than <paramref name="right"/>.
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="left">The lefthand side of the comparison</param>
-        /// <param name="right">The righthand side of the comparison</param>
-        /// <param name="onSuccess">
-        /// The action to take if <paramref name="left"/> is less than <paramref name="right"/>
-        /// </param>
-        /// <param name="onFailure">
-        /// The action to take if <paramref name="left"/> is not less than <paramref name="right"/>
-        /// </param>
-        /// <returns>True if left is less than right.</returns>
-        /// <exception cref="ArgumentNullException">
-        /// If <paramref name="left"/> or <paramref name="right"/> are null.
-        /// </exception>
-        public static bool IsLessThan<T>(T left, T right, Action onSuccess = null, Action onFailure = null)
-        where T : IComparable<T>
-        {
-            EnforceNonNull(left, right);
-            return Passes(() => left.CompareTo(right) < 0, onSuccess, onFailure);
-        }
 
         /// <summary>
         /// Checks if <paramref name="left"/> is equivalent to <paramref name="right"/>.
@@ -301,6 +242,23 @@ namespace Jcd.Utilities.Validations
         {
             EnforceNonNull(left, right);
             return Passes(() => left.CompareTo(right) == 0, onSuccess, onFailure);
+        }
+
+        /// <summary>
+        /// Checks if <paramref name="left"/> and <paramref name="right"/> are the same instance of
+        /// an object.
+        /// </summary>
+        /// <param name="left">The lefthand side of the comparison</param>
+        /// <param name="right">The righthand side of the comparison</param>
+        /// <param name="onSuccess">The action to take if they're the same instance.</param>
+        /// <param name="onFailure">The action to take if they're not the same instance.</param>
+        /// <returns>true if left and right are the same instance.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// If <paramref name="left"/> or <paramref name="right"/> are null.
+        /// </exception>
+        public static bool AreSameObject(object left, object right, Action onSuccess = null, Action onFailure = null)
+        {
+            return Passes(() => ReferenceEquals(left, right), onSuccess, onFailure);
         }
 
         /// <summary>
@@ -333,6 +291,52 @@ namespace Jcd.Utilities.Validations
         {
             EnforceNonNull(value, min, max);
             return Passes(() => min.CompareTo(value) <= 0 && max.CompareTo(value) >= 0, onSuccess, onFailure);
+        }
+
+        /// <summary>
+        /// Checks if <paramref name="left"/> is greater than <paramref name="right"/>.
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="left">The lefthand side of the comparison</param>
+        /// <param name="right">The righthand side of the comparison</param>
+        /// <param name="onSuccess">
+        /// The action to take if <paramref name="left"/> is greater than <paramref name="right"/>
+        /// </param>
+        /// <param name="onFailure">
+        /// The action to take if <paramref name="left"/> is not greater than <paramref name="right"/>
+        /// </param>
+        /// <returns>True if left is greater than right.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// If <paramref name="left"/> or <paramref name="right"/> are null.
+        /// </exception>
+        public static bool IsGreaterThan<T>(T left, T right, Action onSuccess = null, Action onFailure = null)
+      where T : IComparable<T>
+        {
+            EnforceNonNull(left, right);
+            return Passes(() => left.CompareTo(right) > 0, onSuccess, onFailure);
+        }
+
+        /// <summary>
+        /// Checks if <paramref name="left"/> is less than <paramref name="right"/>.
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="left">The lefthand side of the comparison</param>
+        /// <param name="right">The righthand side of the comparison</param>
+        /// <param name="onSuccess">
+        /// The action to take if <paramref name="left"/> is less than <paramref name="right"/>
+        /// </param>
+        /// <param name="onFailure">
+        /// The action to take if <paramref name="left"/> is not less than <paramref name="right"/>
+        /// </param>
+        /// <returns>True if left is less than right.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// If <paramref name="left"/> or <paramref name="right"/> are null.
+        /// </exception>
+        public static bool IsLessThan<T>(T left, T right, Action onSuccess = null, Action onFailure = null)
+        where T : IComparable<T>
+        {
+            EnforceNonNull(left, right);
+            return Passes(() => left.CompareTo(right) < 0, onSuccess, onFailure);
         }
 
         /// <summary>
@@ -373,29 +377,6 @@ namespace Jcd.Utilities.Validations
 
         /// <summary>
         /// Evaluates a predicate ( <paramref name="condition"/> ) on a value ( <paramref
-        /// name="value"/> ) and returns the result of the evaluation.
-        /// </summary>
-        /// <typeparam name="T">The datatype being evaluated</typeparam>
-        /// <param name="condition">The condition being evaluated</param>
-        /// <param name="value">The value being evaluated</param>
-        /// <param name="onSuccess">
-        /// The action to perform if <paramref name="condition"/> returns true when passed <paramref name="value"/>
-        /// </param>
-        /// <param name="onFailure">
-        /// The action to perform if <paramref name="condition"/> returns false when passed <paramref name="value"/>
-        /// </param>
-        /// <returns>
-        /// The result of executing <paramref name="condition"/>( <paramref name="value"/> )
-        /// </returns>
-        /// <exception cref="ArgumentNullException">If <paramref name="condition"/> is null.</exception>
-        public static bool Passes<T>(Signature<T> condition, T value, Action onSuccess = null, Action onFailure = null)
-        {
-            EnforceNonNull(condition);
-            return Passes(() => condition(value), onSuccess, onFailure);
-        }
-
-        /// <summary>
-        /// Evaluates a predicate ( <paramref name="condition"/> ) on a value ( <paramref
         /// name="value"/> ) and returns the logical complement of the result of the evaluation.
         /// </summary>
         /// <typeparam name="T">The datatype being evaluated</typeparam>
@@ -420,36 +401,6 @@ namespace Jcd.Utilities.Validations
 
         /// <summary>
         /// Evaluates a parameterless predicate ( <paramref name="condition"/> ) and returns the
-        /// result of the evaluation.
-        /// </summary>
-        /// <param name="condition">The condition being evaluated</param>
-        /// <param name="onSuccess">
-        /// The action to perform if <paramref name="condition"/> returns true
-        /// </param>
-        /// <param name="onFailure">
-        /// The action to perform if <paramref name="condition"/> returns false
-        /// </param>
-        /// <returns>The result of executing <paramref name="condition"/>()</returns>
-        /// <exception cref="ArgumentNullException">If <paramref name="condition"/> is null.</exception>
-        public static bool Passes(Func<bool> condition, Action onSuccess = null, Action onFailure = null)
-        {
-            EnforceNonNull(condition);
-            var result = condition();
-
-            if (result)
-            {
-                onSuccess?.Invoke();
-            }
-            else
-            {
-                onFailure?.Invoke();
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Evaluates a parameterless predicate ( <paramref name="condition"/> ) and returns the
         /// logical complement of the result of the evaluation.
         /// </summary>
         /// <param name="condition">The condition being evaluated</param>
@@ -465,41 +416,6 @@ namespace Jcd.Utilities.Validations
         {
             EnforceNonNull(condition);
             return Passes(() => !condition(), onSuccess, onFailure);
-        }
-
-        /// <summary>
-        /// Evaluates a set of predicates ( <paramref name="conditions"/> ) on a value to determine
-        /// if all predicates return true.
-        /// </summary>
-        /// <typeparam name="T">The type of data being evaluated</typeparam>
-        /// <param name="conditions">The set of predicates to evaluate</param>
-        /// <param name="value">The value to evaluate the predicates on.</param>
-        /// <param name="onSuccess">The action to perform if all predicates return true.</param>
-        /// <param name="onFailure">The action to perform if any predicate return false.</param>
-        /// <returns>True if all predicates return true</returns>
-        /// <exception cref="ArgumentNullException">
-        /// If <paramref name="conditions"/> is null or any individual entry is null.
-        /// </exception>
-        public static bool PassesAll<T>(IEnumerable<Signature<T>> conditions, T value, Action onSuccess = null,
-                                        Action onFailure = null)
-        {
-            EnforceNonNull(conditions);
-            EnforceAllEntriesNonNull(conditions);
-            bool result = true;
-            int i = 0;
-
-            foreach (var c in conditions)
-            {
-                result = result && (c == null ? result : c(value));
-                i++;
-
-                if (!result)
-                {
-                    break;
-                }
-            }
-
-            return Passes(() => result, onSuccess, onFailure);
         }
 
         /// <summary>
@@ -539,27 +455,6 @@ namespace Jcd.Utilities.Validations
 
         /// <summary>
         /// Evaluates a set of predicates ( <paramref name="conditions"/> ) on a value to determine
-        /// if any predicates return true.
-        /// </summary>
-        /// <typeparam name="T">The type of data being evaluated</typeparam>
-        /// <param name="conditions">The set of predicates to evaluate</param>
-        /// <param name="value">The value to evaluate the predicates on.</param>
-        /// <param name="onSuccess">The action to perform if any predicate return true.</param>
-        /// <param name="onFailure">The action to perform if all predicates return false.</param>
-        /// <returns>True if any predicate return true</returns>
-        /// <exception cref="ArgumentNullException">
-        /// If <paramref name="conditions"/> is null or any individual entry is null.
-        /// </exception>
-        public static bool PassesAny<T>(IEnumerable<Signature<T>> conditions, T value, Action onSuccess = null,
-                                        Action onFailure = null)
-        {
-            EnforceNonNull(conditions);
-            EnforceAllEntriesNonNull(conditions);
-            return !FailsAll(conditions, value, onFailure, onSuccess);
-        }
-
-        /// <summary>
-        /// Evaluates a set of predicates ( <paramref name="conditions"/> ) on a value to determine
         /// if any predicates return false.
         /// </summary>
         /// <typeparam name="T">The type of data being evaluated</typeparam>
@@ -579,7 +474,126 @@ namespace Jcd.Utilities.Validations
             return !PassesAll(conditions, value, onFailure, onSuccess);
         }
 
+        /// <summary>
+        /// Evaluates a predicate ( <paramref name="condition"/> ) on a value ( <paramref
+        /// name="value"/> ) and returns the result of the evaluation.
+        /// </summary>
+        /// <typeparam name="T">The datatype being evaluated</typeparam>
+        /// <param name="condition">The condition being evaluated</param>
+        /// <param name="value">The value being evaluated</param>
+        /// <param name="onSuccess">
+        /// The action to perform if <paramref name="condition"/> returns true when passed <paramref name="value"/>
+        /// </param>
+        /// <param name="onFailure">
+        /// The action to perform if <paramref name="condition"/> returns false when passed <paramref name="value"/>
+        /// </param>
+        /// <returns>
+        /// The result of executing <paramref name="condition"/>( <paramref name="value"/> )
+        /// </returns>
+        /// <exception cref="ArgumentNullException">If <paramref name="condition"/> is null.</exception>
+        public static bool Passes<T>(Signature<T> condition, T value, Action onSuccess = null, Action onFailure = null)
+        {
+            EnforceNonNull(condition);
+            return Passes(() => condition(value), onSuccess, onFailure);
+        }
+
+        /// <summary>
+        /// Evaluates a parameterless predicate ( <paramref name="condition"/> ) and returns the
+        /// result of the evaluation.
+        /// </summary>
+        /// <param name="condition">The condition being evaluated</param>
+        /// <param name="onSuccess">
+        /// The action to perform if <paramref name="condition"/> returns true
+        /// </param>
+        /// <param name="onFailure">
+        /// The action to perform if <paramref name="condition"/> returns false
+        /// </param>
+        /// <returns>The result of executing <paramref name="condition"/>()</returns>
+        /// <exception cref="ArgumentNullException">If <paramref name="condition"/> is null.</exception>
+        public static bool Passes(Func<bool> condition, Action onSuccess = null, Action onFailure = null)
+        {
+            EnforceNonNull(condition);
+            var result = condition();
+
+            if (result)
+            {
+                onSuccess?.Invoke();
+            }
+            else
+            {
+                onFailure?.Invoke();
+            }
+
+            return result;
+        }
+
+        /// <summary>
+        /// Evaluates a set of predicates ( <paramref name="conditions"/> ) on a value to determine
+        /// if all predicates return true.
+        /// </summary>
+        /// <typeparam name="T">The type of data being evaluated</typeparam>
+        /// <param name="conditions">The set of predicates to evaluate</param>
+        /// <param name="value">The value to evaluate the predicates on.</param>
+        /// <param name="onSuccess">The action to perform if all predicates return true.</param>
+        /// <param name="onFailure">The action to perform if any predicate return false.</param>
+        /// <returns>True if all predicates return true</returns>
+        /// <exception cref="ArgumentNullException">
+        /// If <paramref name="conditions"/> is null or any individual entry is null.
+        /// </exception>
+        public static bool PassesAll<T>(IEnumerable<Signature<T>> conditions, T value, Action onSuccess = null,
+                                        Action onFailure = null)
+        {
+            EnforceNonNull(conditions);
+            EnforceAllEntriesNonNull(conditions);
+            bool result = true;
+            int i = 0;
+
+            foreach (var c in conditions)
+            {
+                result = result && (c == null ? result : c(value));
+                i++;
+
+                if (!result)
+                {
+                    break;
+                }
+            }
+
+            return Passes(() => result, onSuccess, onFailure);
+        }
+
+        /// <summary>
+        /// Evaluates a set of predicates ( <paramref name="conditions"/> ) on a value to determine
+        /// if any predicates return true.
+        /// </summary>
+        /// <typeparam name="T">The type of data being evaluated</typeparam>
+        /// <param name="conditions">The set of predicates to evaluate</param>
+        /// <param name="value">The value to evaluate the predicates on.</param>
+        /// <param name="onSuccess">The action to perform if any predicate return true.</param>
+        /// <param name="onFailure">The action to perform if all predicates return false.</param>
+        /// <returns>True if any predicate return true</returns>
+        /// <exception cref="ArgumentNullException">
+        /// If <paramref name="conditions"/> is null or any individual entry is null.
+        /// </exception>
+        public static bool PassesAny<T>(IEnumerable<Signature<T>> conditions, T value, Action onSuccess = null,
+                                        Action onFailure = null)
+        {
+            EnforceNonNull(conditions);
+            EnforceAllEntriesNonNull(conditions);
+            return !FailsAll(conditions, value, onFailure, onSuccess);
+        }
+
         #endregion custom and multi-condition T
+
+        #region Private Methods
+
+        private static void EnforceAllEntriesNonNull<T>(IEnumerable<Signature<T>> conditions)
+        {
+            if (conditions.Any(c => c == null))
+            {
+                throw new ArgumentNullException(nameof(conditions), "All entries must be non-null");
+            }
+        }
 
         private static void EnforceNonNull<T>(params T[] values)
         {
@@ -594,12 +608,6 @@ namespace Jcd.Utilities.Validations
             }
         }
 
-        private static void EnforceAllEntriesNonNull<T>(IEnumerable<Signature<T>> conditions)
-        {
-            if (conditions.Any(c => c == null))
-            {
-                throw new ArgumentNullException(nameof(conditions), "All entries must be non-null");
-            }
-        }
+        #endregion Private Methods
     }
 }
