+++ test/Jcd.Utilities.Test/Formatting/CustomFormatterBaseTests.cs
@@ -7,26 +7,12 @@ namespace Jcd.Utilities.Test.Formatting
 {
     public class CustomFormatterBaseTests
     {
-        public class FakeCustomFormatter : CustomFormatterBase
+        #region Public Methods
+
+        [Fact]
+        public void Constructor_EmptyHandledTypes_ExpectArgumentExceptions()
         {
-            public static Type[] handledTypes = new[] { typeof(int), typeof(Single) };
-            protected string formatResult;
-
-            public FakeCustomFormatter(string formatResult, IEnumerable<Type> handledTypes = null,
-                                       Func<ICustomFormatter, string, object, IFormatProvider, string> formatFunction = null) : base(handledTypes, formatFunction)
-            {
-                this.formatResult = formatResult;
-            }
-
-            public static string Format(ICustomFormatter formatter, string fmt, object arg, IFormatProvider fmtProvider)
-            {
-                if (formatter is FakeCustomFormatter self)
-                {
-                    return self.formatResult;
-                }
-
-                return null;
-            }
+            Assert.Throws<ArgumentException>(() => new FakeCustomFormatter(null, new Type[] { }, null));
         }
 
         [Fact]
@@ -39,12 +25,6 @@ namespace Jcd.Utilities.Test.Formatting
         }
 
         [Fact]
-        public void Constructor_EmptyHandledTypes_ExpectArgumentExceptions()
-        {
-            Assert.Throws<ArgumentException>(() => new FakeCustomFormatter(null, new Type[] { }, null));
-        }
-
-        [Fact]
         public void Format_NullParameters_ExpectArgumentNullException()
         {
             var sut = CreateSut();
@@ -53,9 +33,58 @@ namespace Jcd.Utilities.Test.Formatting
             Assert.Throws<ArgumentNullException>(() => sut.Format(null, new object(), sut));
         }
 
+        #endregion Public Methods
+
+        #region Private Methods
+
         private static FakeCustomFormatter CreateSut()
         {
             return new FakeCustomFormatter("result", FakeCustomFormatter.handledTypes, FakeCustomFormatter.Format);
         }
+
+        #endregion Private Methods
+
+        #region Public Classes
+
+        public class FakeCustomFormatter : CustomFormatterBase
+        {
+            #region Public Fields
+
+            public static Type[] handledTypes = new[] { typeof(int), typeof(Single) };
+
+            #endregion Public Fields
+
+            #region Protected Fields
+
+            protected string formatResult;
+
+            #endregion Protected Fields
+
+            #region Public Constructors
+
+            public FakeCustomFormatter(string formatResult, IEnumerable<Type> handledTypes = null,
+                                    Func<ICustomFormatter, string, object, IFormatProvider, string> formatFunction = null) : base(handledTypes, formatFunction)
+            {
+                this.formatResult = formatResult;
+            }
+
+            #endregion Public Constructors
+
+            #region Public Methods
+
+            public static string Format(ICustomFormatter formatter, string fmt, object arg, IFormatProvider fmtProvider)
+            {
+                if (formatter is FakeCustomFormatter self)
+                {
+                    return self.formatResult;
+                }
+
+                return null;
+            }
+
+            #endregion Public Methods
+        }
+
+        #endregion Public Classes
     }
 }
